+++ doc/source/changelog.rst
@@ -6,6 +6,8 @@ next
 * Boolean flags are now implemented using a variant of
   ``argparse.BooleanOptionalAction``.
 * Added ``no_negated_flags``.
+* A custom parser set for ``type(None)`` now affects parsing of
+  `typing.Optional` too.
 
 6.0.2 (2020-12-08)
 ------------------